Allow `epub:` attributes
At the moment, `epub:` attributes are stripped. There are currently only
two of these attributes defined by [the standard][1]:

  - `epub:type`
  - `epub:prefix`

The `epub:type` prefix in particular is very useful for adding semantic
information about content for things like nicer footnote behaviour. This
change adds the above attributes to the attribute whitelist.

For ease of development, this change also adds `coffeescript` to the
list of development dependencies, and adds an `npm` script so that
the project can be compiled: `npm run compile`.

[1]: http://www.idpf.org/epub/31/spec/epub-contentdocs.html